+++ src/host/asio/mod.rs
@@ -26,7 +26,7 @@ mod stream;
 ///
 /// ASIO only supports loading a single driver at a time globally, so all Host instances
 /// must share the same underlying sys::Asio wrapper to properly coordinate driver access.
-pub(crate) static GLOBAL_ASIO: OnceLock<Arc<sys::Asio>> = OnceLock::new();
+static GLOBAL_ASIO: OnceLock<Arc<sys::Asio>> = OnceLock::new();
 
 /// The host for ASIO.
 #[derive(Debug)]
